+++ app/src/main/java/ru/mail/polis/helpers/GetString.kt
@@ -1,6 +1,6 @@
 package ru.mail.polis.helpers
 
-public fun getAgeString(age: Int): String {
+public fun getAgeString(age: Long): String {
     return when {
         age % 100 in 5..20 -> {
             "$age лет"
@@ -8,7 +8,7 @@ public fun getAgeString(age: Int): String {
         age % 10 in 2..4 -> {
             "$age года"
         }
-        age % 10 == 1 -> {
+        age % 10 == 1L -> {
             "$age год"
         }
         else -> {
